backend/requirements.txt
fastapi
uvicorn[standard]
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
langchain-google-genai>=2.0.0
faiss-cpu
sentence-transformers
pypdf
python-dotenv
pydantic

